Break input handler into its own file
