+++ theme/spriteeditor.less
@@ -267,3 +267,12 @@
         background: rgb(44, 62, 80);
     }
 }
+
+.ui.modal.fullscreen.sprite-editor-snippet-portal {
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    z-index: 9999;
+}
